# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email & Password does not match with our record."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "description": "Register by email, username, password",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Pass user register details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created Successfully"
                                        },
                                        "user": {
                                            "type": "string",
                                            "example": "1|PFx96rYVBPzZS3Thuw0Bak6KugaaSMiWCq4imusv"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books": {
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Cerate Book",
                "operationId": "boookCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "topic",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "موش و گربه"
                                    },
                                    "topic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "کودکان"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "خذتثطف",
                                        "example": "نوشته نشده"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "موش و گربه"
                                        },
                                        "topic": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "کودکان"
                                        },
                                        "description": {
                                            "type": "string",
                                            "format": "خذتثطف",
                                            "example": "نوشته نشده"
                                        },
                                        "author_id": {
                                            "type": "integer",
                                            "format": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "put": {
                "tags": [
                    "books"
                ],
                "summary": "Cerate Book",
                "operationId": "boookUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "topic",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "موش و گربه"
                                    },
                                    "topic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "کودکان"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "نوشته نشده"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "کتاب با موفقیت آپدیت شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "books"
                ],
                "summary": "Delete Book",
                "operationId": "bookDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "کتاب با موفقیت حذف شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}